@@ -5,16 +5,15 @@ author: Rex Isaac Raphael
 homepage: https://github.com/juicycleff/sanity_dart_client
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
-  flutter:
-    sdk: flutter
-
+  json_annotation: ^3.1.1
   dio: ^3.0.10
   rxdart: ^0.24.1
 
 dev_dependencies:
+  build_runner:
   effective_dart: ^1.3.0
   test: ^1.15.7
+  json_serializable: ^3.5.1
