@@ -3,22 +3,22 @@ part of 'package:sanity_client_flutter/sanity_client_flutter.dart';
 /// A simple crud interface
 abstract class Crud<C, R, U, D> {
   /// Create an entity
-  Future<R> create(C data) {
+  Future<R?>? create(C data) {
     return null;
   }
 
   /// Update an entity
-  Future<R> update(U data) {
+  Future<R?>? update(U data) {
     return null;
   }
 
   /// Delete an entity
-  Future<D> delete(String id) {
+  Future<D?>? delete(String id) {
     return null;
   }
 
   /// List all entity
-  Future<Iterable<R>> list() {
+  Future<Iterable<R?>>? list() {
     return null;
   }
 }
@@ -26,7 +26,7 @@ abstract class Crud<C, R, U, D> {
 /// A simple crud interface
 abstract class CrudWithGet<C, R, U, D> extends Crud<C, R, U, D> {
   /// Get an entity
-  Future<R> getOne(String id) {
+  Future<R?>? getOne(String id) {
     return null;
   }
 }
@@ -34,8 +34,8 @@ abstract class CrudWithGet<C, R, U, D> extends Crud<C, R, U, D> {
 /// Request option
 abstract class RequestOptions {
   /// Request timeout
-  int timeout;
+  int? timeout;
 
   /// request token
-  String token;
+  String? token;
 }
